code/backend/requirements.txt
# ── Core Framework ────────────────────────────────────────────────────────────
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# ── Database ──────────────────────────────────────────────────────────────────
SQLAlchemy==2.0.40
alembic==1.15.2
psycopg2-binary==2.9.10

# ── Authentication & Security ─────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
cryptography==44.0.2
pyotp==2.9.0
qrcode[pil]==8.1

# ── Configuration ─────────────────────────────────────────────────────────────
pydantic==2.11.3
pydantic-settings==2.8.1
python-dotenv==1.1.0

# ── Redis & Caching ───────────────────────────────────────────────────────────
redis==5.2.1
hiredis==3.1.1

# ── HTTP & API ────────────────────────────────────────────────────────────────
requests==2.32.3
httpx==0.28.1
aiohttp==3.11.18
aiofiles==24.1.0

# ── Data Processing ───────────────────────────────────────────────────────────
pandas==2.2.3
numpy==2.2.4
scipy==1.15.2

# ── Machine Learning ──────────────────────────────────────────────────────────
scikit-learn==1.6.1
joblib==1.4.2
xgboost==2.1.4

# ── Time-Series & Forecasting ────────────────────────────────────────────────
statsmodels==0.14.4
arch==7.2.0

# ── Portfolio Optimisation ───────────────────────────────────────────────────
cvxpy==1.6.4

# ── Financial Data ────────────────────────────────────────────────────────────
yfinance==0.2.54
alpha-vantage==3.0.0

# ── Blockchain / Web3 ────────────────────────────────────────────────────────
web3==7.10.0

# ── NLP / Sentiment ──────────────────────────────────────────────────────────
openai==1.75.0

# ── Logging & Monitoring ─────────────────────────────────────────────────────
structlog==25.3.0
sentry-sdk[fastapi]==2.25.1
psutil==7.0.0
prometheus-client==0.21.1

# ── File Handling ─────────────────────────────────────────────────────────────
openpyxl==3.1.5
Pillow==11.2.1

# ── Async / Background Tasks ─────────────────────────────────────────────────
APScheduler==3.11.0

# ── Utilities ────────────────────────────────────────────────────────────────
python-dateutil==2.9.0.post0
pytz==2025.2
click==8.1.8

# ── Production Server ────────────────────────────────────────────────────────
gunicorn==23.0.0

# ── Testing ──────────────────────────────────────────────────────────────────
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.0
pytest-mock==3.14.0
faker==37.1.0
httpx==0.28.1
